@@ -53,7 +53,7 @@
   "private": true,
   "dependencies": {
     "@angular/animations": "10.2.4",
-    "@angular/cdk": "11.2.7",
+    "@angular/cdk": "11.2.11",
     "@angular/common": "10.2.4",
     "@angular/compiler": "10.2.4",
     "@angular/core": "10.2.4",
@@ -71,7 +71,6 @@
     "@fullcalendar/interaction": "5.6.0",
     "@fullcalendar/list": "5.6.0",
     "@fullcalendar/timegrid": "5.6.0",
-    "@neuronet.io/vido": "2.6.22",
     "@ng-select/ng-option-highlight": "0.0.6",
     "@ng-select/ng-select": "6.1.0",
     "@ngneat/dialog": "1.4.0",
@@ -79,18 +78,18 @@
     "@ngneat/reactive-forms": "1.7.2",
     "@ngneat/transloco": "2.20.1",
     "@ngneat/until-destroy": "8.0.4",
-    "@ngx-loading-bar/core": "5.1.1",
-    "@ngx-loading-bar/http-client": "5.1.1",
-    "@ngx-loading-bar/router": "5.1.1",
+    "@ngx-loading-bar/core": "5.1.2",
+    "@ngx-loading-bar/http-client": "5.1.2",
+    "@ngx-loading-bar/router": "5.1.2",
     "@nrwl/angular": "11.5.2",
     "@tinymce/tinymce-angular": "4.2.2",
     "angular-gridster2": "11.1.5",
     "angular-resizable-element": "3.3.5",
     "bootstrap": "4.6.0",
     "cropperjs": "1.5.11",
-    "d3": "6.6.2",
-    "date-fns": "2.19.0",
-    "filesize": "6.1.0",
+    "d3": "6.7.0",
+    "date-fns": "2.21.1",
+    "filesize": "6.3.0",
     "flatpickr": "^4.6.9",
     "hammerjs": "2.0.8",
     "lodash-es": "4.17.21",
@@ -122,15 +121,15 @@
     "@angular/cli": "10.2.1",
     "@angular/compiler-cli": "10.2.4",
     "@angular/language-service": "10.2.4",
-    "@babel/core": "7.13.14",
-    "@babel/plugin-transform-runtime": "7.13.10",
-    "@babel/preset-env": "7.13.12",
+    "@babel/core": "7.13.16",
+    "@babel/plugin-transform-runtime": "7.13.15",
+    "@babel/preset-env": "7.13.15",
     "@commitlint/cli": "12.1.1",
     "@datorama/akita-cli": "6.1.1",
     "@datorama/akita-ng-router-store": "6.0.0",
     "@datorama/akita-ngdevtools": "6.0.0",
     "@ngneat/spectator": "7.1.0",
-    "@ngneat/transloco-keys-manager": "2.7.1",
+    "@ngneat/transloco-keys-manager": "2.7.2",
     "@ngneat/transloco-optimize": "^1.0.2",
     "@nrwl/cli": "11.5.2",
     "@nrwl/cypress": "11.5.2",
@@ -141,14 +140,14 @@
     "@schematics/angular": "10.2.1",
     "@types/chance": "1.1.1",
     "@types/d3": "6.3.0",
-    "@types/jest": "26.0.22",
+    "@types/jest": "26.0.23",
     "@types/lodash-es": "4.17.4",
     "@types/mime": "^2.0.3",
     "@types/node": "14.14.37",
     "@types/uuid": "8.3.0",
-    "@typescript-eslint/eslint-plugin": "4.21.0",
-    "@typescript-eslint/eslint-plugin-tslint": "4.21.0",
-    "@typescript-eslint/parser": "4.21.0",
+    "@typescript-eslint/eslint-plugin": "4.22.0",
+    "@typescript-eslint/eslint-plugin-tslint": "4.22.0",
+    "@typescript-eslint/parser": "4.22.0",
     "babel-jest": "26.6.3",
     "chance": "1.1.7",
     "codelyzer": "6.0.1",
@@ -156,10 +155,10 @@
     "commitlint-plugin-jira-rules": "1.5.1",
     "cypress": "6.8.0",
     "dotenv": "8.2.0",
-    "eslint": "7.23.0",
-    "eslint-config-prettier": "8.1.0",
+    "eslint": "7.25.0",
+    "eslint-config-prettier": "8.3.0",
     "eslint-plugin-cypress": "2.11.2",
-    "eslint-plugin-prettier": "3.3.1",
+    "eslint-plugin-prettier": "3.4.0",
     "fast-check": "2.14.0",
     "husky": "4.3.8",
     "jest": "26.6.3",
@@ -168,7 +167,7 @@
     "ng-packagr": "10.1.2",
     "patch-package": "6.4.7",
     "prettier": "2.2.1",
-    "ts-jest": "26.5.4",
+    "ts-jest": "26.5.5",
     "ts-node": "9.1.1",
     "tsickle": "0.39.1",
     "tslint": "6.1.3",
