@@ -8,8 +8,10 @@ import { ChangeDetectionStrategy, ChangeDetectorRef, Component, EventEmitter, On
 import { Validators } from '@angular/forms';
 import { Title } from '@angular/platform-browser';
 import { ActivatedRoute, Router } from '@angular/router';
+import { ProjectSidebarItem } from '@app/enums/project-sidebar-item.enum';
 import { NewAppointmentModalComponent } from '@app/modules/appointment/components/modals/new/new.component';
 import { PendingChangesModalComponent } from '@app/modules/shared/modals/pending-changes/pending-changes.component';
+import { LeaveProjectModalComponent } from '@app/pages/projects/components/modals/leave/leave.component';
 import {
   AppointmentsService,
   AuthService,
@@ -19,7 +21,7 @@ import {
   SearchService,
   WebSocketService,
 } from '@app/services';
-import { UserService } from '@app/stores/user';
+import { UserService, UserStore } from '@app/stores/user';
 import {
   Appointment,
   AppointmentPayload,
@@ -68,6 +70,8 @@ export class AppointmentPageComponent implements OnInit, OnDestroy {
 
   public id = this.route.snapshot.paramMap.get('id')!;
 
+  public sidebarItem = ProjectSidebarItem.Appointments;
+
   public currentUser: User | null = null;
 
   public initialState?: Appointment;
@@ -95,6 +99,8 @@ export class AppointmentPageComponent implements OnInit, OnDestroy {
   public lock: Lock | null = null;
 
   public modified = false;
+
+  public showSidebar = false;
 
   public loading = true;
 
@@ -142,7 +148,8 @@ export class AppointmentPageComponent implements OnInit, OnDestroy {
     private readonly searchService: SearchService,
     private readonly pageTitleService: PageTitleService,
     private readonly titleService: Title,
-    private readonly modalService: DialogService
+    private readonly modalService: DialogService,
+    private readonly userStore: UserStore
   ) {}
 
   public get f(): FormGroup<FormAppointment>['controls'] {
@@ -240,6 +247,7 @@ export class AppointmentPageComponent implements OnInit, OnDestroy {
     );
 
     this.initTranslations();
+    this.initSidebar();
     this.initSearchInput();
     this.initDetails();
     this.initPageTitle();
@@ -278,6 +286,20 @@ export class AppointmentPageComponent implements OnInit, OnDestroy {
         })
       )
       .subscribe();
+  }
+
+  public initSidebar(): void {
+    this.route.params.subscribe(params => {
+      if (params.projectId) {
+        this.showSidebar = true;
+
+        this.projectsService.get(params.projectId).subscribe(
+          /* istanbul ignore next */ project => {
+            this.projects = [...this.projects, project];
+          }
+        );
+      }
+    });
   }
 
   public initSearchInput(): void {
@@ -485,6 +507,32 @@ export class AppointmentPageComponent implements OnInit, OnDestroy {
           this.cdr.markForCheck();
         }
       );
+  }
+
+  public canDeactivate(): Observable<boolean> {
+    if (this.showSidebar) {
+      const userStoreValue = this.userStore.getValue();
+      const userSetting = 'SkipDialog-LeaveProject';
+
+      /* istanbul ignore next */
+      const skipLeaveDialog = Boolean(userStoreValue.user?.userprofile.ui_settings?.confirm_dialog?.[userSetting]);
+
+      if (skipLeaveDialog) {
+        return of(true);
+      }
+
+      this.modalRef = this.modalService.open(LeaveProjectModalComponent, {
+        closeButton: false,
+      });
+      /* istanbul ignore next */
+      return this.modalRef.afterClosed$.pipe(
+        untilDestroyed(this),
+        take(1),
+        map(val => Boolean(val))
+      );
+    }
+
+    return of(true);
   }
 
   public pendingChanges(): Observable<boolean> {
