@@ -8,8 +8,11 @@ import { ChangeDetectionStrategy, ChangeDetectorRef, Component, EventEmitter, On
 import { Validators } from '@angular/forms';
 import { Title } from '@angular/platform-browser';
 import { ActivatedRoute, Router } from '@angular/router';
+import { ProjectSidebarItem } from '@app/enums/project-sidebar-item.enum';
 import { PendingChangesModalComponent } from '@app/modules/shared/modals/pending-changes/pending-changes.component';
+import { LeaveProjectModalComponent } from '@app/pages/projects/components/modals/leave/leave.component';
 import { AuthService, ContactsService, PageTitleService, ProjectsService, WebSocketService } from '@app/services';
+import { UserStore } from '@app/stores/user';
 import { Contact, ContactPayload, Lock, Metadata, Privileges, Project, User } from '@eworkbench/types';
 import { DialogRef, DialogService } from '@ngneat/dialog';
 import { FormBuilder, FormGroup } from '@ngneat/reactive-forms';
@@ -43,6 +46,8 @@ export class ContactPageComponent implements OnInit, OnDestroy {
 
   public id = this.route.snapshot.paramMap.get('id')!;
 
+  public sidebarItem = ProjectSidebarItem.Contacts;
+
   public currentUser: User | null = null;
 
   public initialState?: Contact;
@@ -54,6 +59,8 @@ export class ContactPageComponent implements OnInit, OnDestroy {
   public lock: Lock | null = null;
 
   public modified = false;
+
+  public showSidebar = false;
 
   public loading = true;
 
@@ -97,7 +104,8 @@ export class ContactPageComponent implements OnInit, OnDestroy {
     private readonly projectsService: ProjectsService,
     private readonly pageTitleService: PageTitleService,
     private readonly titleService: Title,
-    private readonly modalService: DialogService
+    private readonly modalService: DialogService,
+    private readonly userStore: UserStore
   ) {}
 
   public get f(): FormGroup<FormContact>['controls'] {
@@ -161,6 +169,7 @@ export class ContactPageComponent implements OnInit, OnDestroy {
       }
     );
 
+    this.initSidebar();
     this.initSearchInput();
     this.initDetails();
     this.initPageTitle();
@@ -184,6 +193,20 @@ export class ContactPageComponent implements OnInit, OnDestroy {
         })
       )
       .subscribe();
+  }
+
+  public initSidebar(): void {
+    this.route.params.subscribe(params => {
+      if (params.projectId) {
+        this.showSidebar = true;
+
+        this.projectsService.get(params.projectId).subscribe(
+          /* istanbul ignore next */ project => {
+            this.projects = [...this.projects, project];
+          }
+        );
+      }
+    });
   }
 
   public initSearchInput(): void {
@@ -338,6 +361,32 @@ export class ContactPageComponent implements OnInit, OnDestroy {
           this.cdr.markForCheck();
         }
       );
+  }
+
+  public canDeactivate(): Observable<boolean> {
+    if (this.showSidebar) {
+      const userStoreValue = this.userStore.getValue();
+      const userSetting = 'SkipDialog-LeaveProject';
+
+      /* istanbul ignore next */
+      const skipLeaveDialog = Boolean(userStoreValue.user?.userprofile.ui_settings?.confirm_dialog?.[userSetting]);
+
+      if (skipLeaveDialog) {
+        return of(true);
+      }
+
+      this.modalRef = this.modalService.open(LeaveProjectModalComponent, {
+        closeButton: false,
+      });
+      /* istanbul ignore next */
+      return this.modalRef.afterClosed$.pipe(
+        untilDestroyed(this),
+        take(1),
+        map(val => Boolean(val))
+      );
+    }
+
+    return of(true);
   }
 
   public pendingChanges(): Observable<boolean> {
