@@ -18,10 +18,12 @@ import { Validators } from '@angular/forms';
 import { Title } from '@angular/platform-browser';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ModalState } from '@app/enums/modal-state.enum';
+import { ProjectSidebarItem } from '@app/enums/project-sidebar-item.enum';
 import { RemoveResourcePDFModalComponent } from '@app/modules/resource/components/modals/remove-pdf/remove-pdf.component';
 import { PendingChangesModalComponent } from '@app/modules/shared/modals/pending-changes/pending-changes.component';
+import { LeaveProjectModalComponent } from '@app/pages/projects/components/modals/leave/leave.component';
 import { AuthService, PageTitleService, ProjectsService, ResourcesService, UserGroupsService, WebSocketService } from '@app/services';
-import { UserService } from '@app/stores/user';
+import { UserService, UserStore } from '@app/stores/user';
 import {
   BookingRulesPayload,
   DropdownElement,
@@ -73,6 +75,8 @@ export class ResourcePageComponent implements OnInit, OnDestroy {
 
   public id = this.route.snapshot.paramMap.get('id')!;
 
+  public sidebarItem = ProjectSidebarItem.Resources;
+
   public currentUser: User | null = null;
 
   public initialState?: Resource;
@@ -90,6 +94,8 @@ export class ResourcePageComponent implements OnInit, OnDestroy {
   public lock: Lock | null = null;
 
   public modified = false;
+
+  public showSidebar = false;
 
   public loading = true;
 
@@ -151,7 +157,8 @@ export class ResourcePageComponent implements OnInit, OnDestroy {
     private readonly projectsService: ProjectsService,
     private readonly userGroupsService: UserGroupsService,
     private readonly pageTitleService: PageTitleService,
-    private readonly titleService: Title
+    private readonly titleService: Title,
+    private readonly userStore: UserStore
   ) {}
 
   public get f(): FormGroup<FormResource>['controls'] {
@@ -230,6 +237,7 @@ export class ResourcePageComponent implements OnInit, OnDestroy {
     );
 
     this.initTranslations();
+    this.initSidebar();
     this.initSearchInput();
     this.initDetails();
     this.initPageTitle();
@@ -295,6 +303,20 @@ export class ResourcePageComponent implements OnInit, OnDestroy {
         })
       )
       .subscribe();
+  }
+
+  public initSidebar(): void {
+    this.route.params.subscribe(params => {
+      if (params.projectId) {
+        this.showSidebar = true;
+
+        this.projectsService.get(params.projectId).subscribe(
+          /* istanbul ignore next */ project => {
+            this.projects = [...this.projects, project];
+          }
+        );
+      }
+    });
   }
 
   public initSearchInput(): void {
@@ -502,6 +524,32 @@ export class ResourcePageComponent implements OnInit, OnDestroy {
           this.cdr.markForCheck();
         }
       );
+  }
+
+  public canDeactivate(): Observable<boolean> {
+    if (this.showSidebar) {
+      const userStoreValue = this.userStore.getValue();
+      const userSetting = 'SkipDialog-LeaveProject';
+
+      /* istanbul ignore next */
+      const skipLeaveDialog = Boolean(userStoreValue.user?.userprofile.ui_settings?.confirm_dialog?.[userSetting]);
+
+      if (skipLeaveDialog) {
+        return of(true);
+      }
+
+      this.modalRef = this.modalService.open(LeaveProjectModalComponent, {
+        closeButton: false,
+      });
+      /* istanbul ignore next */
+      return this.modalRef.afterClosed$.pipe(
+        untilDestroyed(this),
+        take(1),
+        map(val => Boolean(val))
+      );
+    }
+
+    return of(true);
   }
 
   public pendingChanges(): Observable<boolean> {
