@@ -10,7 +10,6 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { ModalState } from '@app/enums/modal-state.enum';
 import { ProjectSidebarItem } from '@app/enums/project-sidebar-item.enum';
 import { NewAppointmentModalComponent } from '@app/modules/appointment/components/modals/new/new.component';
-import { RestoreModalComponent } from '@app/modules/trash/components/modals/restore/restore.component';
 import { LeaveProjectModalComponent } from '@app/pages/projects/components/modals/leave/leave.component';
 import { AuthService, PageTitleService, ProjectsService, ResourcesService } from '@app/services';
 import { UserService, UserStore } from '@app/stores/user';
@@ -100,8 +99,6 @@ export class ResourcesPageComponent implements OnInit {
   public project?: string;
 
   public sorting?: TableSortChangedEvent;
-
-  public restoreEmitter = new EventEmitter<ModalCallback>();
 
   public userAvailabilityChoices: DropdownElement[] = [];
 
@@ -374,6 +371,21 @@ export class ResourcesPageComponent implements OnInit {
       });
   }
 
+  public onFilterItems(showTrashedItems: boolean): void {
+    if (showTrashedItems) {
+      this.params = this.params.set('deleted', 'true');
+    } else {
+      this.params = this.params.delete('deleted');
+    }
+    this.tableView.loadData(false, this.params);
+  }
+
+  public onRestore(restored: boolean): void {
+    if (restored) {
+      this.tableView.loadData(false, this.params);
+    }
+  }
+
   public onColumnsChanged(event: TableColumnChangedEvent): void {
     const merged = merge(
       keyBy(event, 'key'),
@@ -470,29 +482,17 @@ export class ResourcesPageComponent implements OnInit {
     /* istanbul ignore next */
     this.modalRef = this.modalService.open(NewResourceModalComponent, {
       closeButton: false,
-      data: { initialState: { projects: this.project ? [this.project] : [] } },
+      data: { withSidebar: this.showSidebar, initialState: { projects: this.project ? [this.project] : [] } },
     });
     /* istanbul ignore next */
     this.modalRef.afterClosed$.pipe(untilDestroyed(this), take(1)).subscribe((callback: ModalCallback) => this.onModalClose(callback));
-  }
-
-  public openTrashModal(): void {
-    /* istanbul ignore next */
-    const subscription = this.restoreEmitter.pipe(untilDestroyed(this)).subscribe((callback: ModalCallback) => this.onModalClose(callback));
-    /* istanbul ignore next */
-    this.modalRef = this.modalService.open(RestoreModalComponent, {
-      closeButton: false,
-      data: { service: this.resourcesService, routerBaseLink: '/resources', restoreEmitter: this.restoreEmitter },
-    });
-    /* istanbul ignore next */
-    this.modalRef.afterClosed$.pipe(untilDestroyed(this), take(1)).subscribe(() => subscription.unsubscribe());
   }
 
   public onOpenDuplicateModal(resource: Resource): void {
     /* istanbul ignore next */
     this.modalRef = this.modalService.open(NewResourceModalComponent, {
       closeButton: false,
-      data: { initialState: resource },
+      data: { withSidebar: this.showSidebar, initialState: resource },
     });
     /* istanbul ignore next */
     this.modalRef.afterClosed$.pipe(untilDestroyed(this), take(1)).subscribe((callback: ModalCallback) => this.onModalClose(callback));
@@ -513,7 +513,7 @@ export class ResourcesPageComponent implements OnInit {
 
   public onModalClose(callback?: ModalCallback): void {
     if (callback?.navigate) {
-      this.router.navigate(callback.navigate);
+      this.router.navigate(callback.navigate, { relativeTo: this.route });
     } else if (callback?.state === ModalState.Changed) {
       this.tableView.loadData();
     }
