@@ -9,11 +9,13 @@ import { Validators } from '@angular/forms';
 import { Title } from '@angular/platform-browser';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ModalState } from '@app/enums/modal-state.enum';
+import { ProjectSidebarItem } from '@app/enums/project-sidebar-item.enum';
 import { NewLabelModalComponent } from '@app/modules/label/component/modals/new/new.component';
 import { PendingChangesModalComponent } from '@app/modules/shared/modals/pending-changes/pending-changes.component';
 import { NewTaskModalComponent } from '@app/modules/task/components/modals/new/new.component';
+import { LeaveProjectModalComponent } from '@app/pages/projects/components/modals/leave/leave.component';
 import { AuthService, LabelsService, PageTitleService, ProjectsService, TasksService, WebSocketService } from '@app/services';
-import { UserService } from '@app/stores/user';
+import { UserService, UserStore } from '@app/stores/user';
 import {
   DateGroup,
   DropdownElement,
@@ -61,6 +63,8 @@ export class TaskPageComponent implements OnInit, OnDestroy {
 
   public id = this.route.snapshot.paramMap.get('id')!;
 
+  public sidebarItem = ProjectSidebarItem.Tasks;
+
   public currentUser: User | null = null;
 
   public initialState?: Task;
@@ -72,6 +76,8 @@ export class TaskPageComponent implements OnInit, OnDestroy {
   public lock: Lock | null = null;
 
   public modified = false;
+
+  public showSidebar = false;
 
   public loading = true;
 
@@ -130,7 +136,8 @@ export class TaskPageComponent implements OnInit, OnDestroy {
     private readonly userService: UserService,
     private readonly labelsService: LabelsService,
     private readonly pageTitleService: PageTitleService,
-    private readonly titleService: Title
+    private readonly titleService: Title,
+    private readonly userStore: UserStore
   ) {}
 
   public get f(): FormGroup<FormTask>['controls'] {
@@ -217,6 +224,7 @@ export class TaskPageComponent implements OnInit, OnDestroy {
     );
 
     this.initTranslations();
+    this.initSidebar();
     this.initSearchInput();
     this.initDetails();
     this.initPageTitle();
@@ -264,6 +272,20 @@ export class TaskPageComponent implements OnInit, OnDestroy {
         debounceTime(500)
       )
       .subscribe();
+  }
+
+  public initSidebar(): void {
+    this.route.params.subscribe(params => {
+      if (params.projectId) {
+        this.showSidebar = true;
+
+        this.projectsService.get(params.projectId).subscribe(
+          /* istanbul ignore next */ project => {
+            this.projects = [...this.projects, project];
+          }
+        );
+      }
+    });
   }
 
   public initSearchInput(): void {
@@ -472,6 +494,32 @@ export class TaskPageComponent implements OnInit, OnDestroy {
           this.cdr.markForCheck();
         }
       );
+  }
+
+  public canDeactivate(): Observable<boolean> {
+    if (this.showSidebar) {
+      const userStoreValue = this.userStore.getValue();
+      const userSetting = 'SkipDialog-LeaveProject';
+
+      /* istanbul ignore next */
+      const skipLeaveDialog = Boolean(userStoreValue.user?.userprofile.ui_settings?.confirm_dialog?.[userSetting]);
+
+      if (skipLeaveDialog) {
+        return of(true);
+      }
+
+      this.modalRef = this.modalService.open(LeaveProjectModalComponent, {
+        closeButton: false,
+      });
+      /* istanbul ignore next */
+      return this.modalRef.afterClosed$.pipe(
+        untilDestroyed(this),
+        take(1),
+        map(val => Boolean(val))
+      );
+    }
+
+    return of(true);
   }
 
   public pendingChanges(): Observable<boolean> {
