@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
+import { ContentTypeModels } from './content-type-model.interface';
 import { User } from './user.interface';
 
 export interface Notification {
   content_type: number;
-  content_type_model: string;
+  content_type_model: ContentTypeModels;
   created_at: string;
   created_by: User;
   display: string;
